requirements.txt
pymongo~=3.11.3
python-dotenv~=0.15.0
pytest~=6.2.2
flask~=1.1.2
ariadne~=0.13.0
werkzeug~=1.0.1
pandas~=1.2.4
requests~=2.25.1


